+++ packages/moonstone/VirtualList/useSpotlightConfig.js
@@ -20,18 +20,18 @@ const useSpotlightConfig = (instance, props) => {
         Spotlight.set(spotlightId, {
             enterTo: 'last-focused',
             /*
-                * Returns the data-index as the key for last focused
-                */
+             * Returns the data-index as the key for last focused
+             */
             lastFocusedPersist: lastFocusedPersist,
             /*
-                * Restores the data-index into the placeholder if its the only element. Tries to find a
-                * matching child otherwise.
-                */
+             * Restores the data-index into the placeholder if its the only element. Tries to find a
+             * matching child otherwise.
+             */
             lastFocusedRestore: lastFocusedRestore,
             /*
-                * Directs spotlight focus to favor straight elements that are within range of `spacing`
-                * over oblique elements, like scroll buttons.
-                */
+             * Directs spotlight focus to favor straight elements that are within range of `spacing`
+             * over oblique elements, like scroll buttons.
+             */
             obliqueMultiplier: spacing > 0 ? spacing : 1
         });
     }
@@ -47,9 +47,9 @@ const useSpotlightConfig = (instance, props) => {
     }
 
     /*
-        * Restores the data-index into the placeholder if it exists. Tries to find a matching child
-        * otherwise.
-        */
+     * Restores the data-index into the placeholder if it exists. Tries to find a matching child
+     * otherwise.
+     */
     function lastFocusedRestore ({key}, all) {
         const placeholder = all.find(el => 'vlPlaceholder' in el.dataset);
         if (placeholder) {
